Added empty solution.
It's named AutoCheck, after QuickCheck. Though, the plan isn't to create
just another QuickCheck clone.

AutoCheck is going to be based on QuickCheck version 1.2.0.1, the last
version of QuickCheck depending on base (prelude and its support libraries)
and Random.

Random is a basic random number generation library, including the ability
to split random number generators. Actually, QuickCheck 1.2.0.1 uses only
Random, StdGen, newStdGen, randomR, and split, from the Random package,
and so it'll be easier to start by porting those in .NET/F#.

The other thing to notice about QuickCheck version 1.2.0.1, is that it's
the last version before version 2 came out. In version 2.0 there's added
support for shrinking, which is a major advancement. Instead, AutoCheck
should try to adopt SmartCheck's smart shrinking approach.

One other difference with QuickCheck version 2 (but also version 1) is that
AutoCheck should support writing properties using existentials as well as
universals. For this, AutoCheck should try to adopt SmallCheck's approach
on that.