Ported Gen.generate.
The signature of the equivalent function in QuickCheck 1.2.0.1 is:

    generate :: Int -> StdGen -> Gen a -> a

The signature of this (ported) function though is:

    val generate : n:'a -> Gen<'b> -> 'b

Notice that it takes one less argument; instead of taking StdGen it creates
an StdGen internally, because the whole Random module which contains StdGen
is marked as internal.

Besides the fact that the Random module is internal (and so there's no way
to take StdGen as an argument in Gen.generate), another reason that StdGen
is created internally is to make Gen.generate truly random.

Let's look at Haskell's Random for a second:

When you call mkStdGen with seeds that are close together you will get two
'similar' generators. For example, the following call produces always the
same output:

    λ> random (mkStdGen 100) :: (Bool, StdGen)
    (True,4041414 40692)

    λ> random (mkStdGen 100) :: (Bool, StdGen)
    (True,4041414 40692)

    λ> random (mkStdGen 100) :: (Bool, StdGen)
    (True,4041414 40692)

And so, in QuickCheck 1.2.0.1, the following calls to the generate function
will produce the same output:

    λ> generate 10 (mkStdGen 70) arbitrary :: [Int]
    [-9,-5,5,5,-3]

    λ> generate 10 (mkStdGen 70) arbitrary :: [Int]
    [-9,-5,5,5,-3]

    λ> generate 10 (mkStdGen 70) arbitrary :: [Int]
    [-9,-5,5,5,-3]

Clearly, taking an StdGen as an argument in generate doesn't produce the
expected (randomized) output.

For this reason, the ported function in F# creates a StdGen with a random
seed:

    let randomSeed = int32 DateTime.UtcNow.Ticks

    let generate n (Gen m) =
        let rnd = createStdGen randomSeed
        let (size, rnd') = randomR (0, n) rnd
        m size rnd'

This might get optimized or changed future commits, but looks fine for now.