Ported Gen.choose.
The signature of the equivalent function in QuickCheck 1.2.0.1 is:

    choose :: Random a => (a, a) -> Gen a

The signature of this (ported) function though is:

    val choose : 'a * 'b -> Gen<'a>

Notice the generic type parameter 'b. This might cause issues down the road
but for now I'm not going to add a type annotation; instead I will leave it
as-is and keep an eye on what might happens.

Another thing worth keeping track of is the two implementations, where in
QuickCheck 1.2.0.1 the function is implemented as:

    choose :: Random a => (a, a) -> Gen a
    choose bounds = (fst . randomR bounds) `fmap` rand
      where rand = Gen (\n r -> r)

In F#, given that we already defined map, and <!>, we could write it as:

    let choose bounds =
        let rand = Gen (fun n r -> r)
        (fst << randomR bounds) <!> rand

I believe that F# has already many infix operators available, and using yet
another, custom, one can make the code harder to understand, thus I decided
to implement the function as:

    let choose bounds =
        Gen (fun n r -> r) |> map (fst << randomR bounds)

Still, I think that the backward composition operator (<<) is not idiomatic
in F#, but at least it corresponds to Haskell's composition operator (.).