Provided more descriptive names for the functions in the Random module, and also used `Random.` on each call from the Gen module.
One might argue that, since this is a port, it might be difficult to go
back and cross-check the original and ported implementation because the
function names are different, and I'd agree.

That won't be the case soon because in future commits there's going to be
documentation for each function, and the documentation will refer to its
original name in Haskell's System.Random package.