Made Gen.generate referencial transparent.
In commit 99e0c98eafccb6f55c252b9db460e21380bf9785 there is a message that
shows Gen.generate's original signature in Haskell QuickCheck 1.2.0.1:

    generate :: Int -> StdGen -> Gen a -> a

Notice the StdGen parameter there. It's for the same reason, to make the
function referencial transparent. Calling Gen.generate with the same params
yields the same result back.

That wasn't the case here, as long as the seed value was calculated in the
function. For this reason, the seed is now a parameter.

It's up to the caller now, when he or she supplied a different seed or not,
the results are the different or same, respectively.

That's probably the reason why Gen.generate' signature and behaviour is
different in QuickCheck 2.